1. Added class StringCrypter. Removed Desipher 2. Updated working with keys

+++ src/main/java/Frames/LicenseWindow.java
@@ -11,10 +11,7 @@ import javax.swing.*;
 import java.awt.*;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
-import java.io.ByteArrayInputStream;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.UnsupportedEncodingException;
+import java.io.*;
 
 public class LicenseWindow extends JFrame {
     private JLabel textLabel = new JLabel("<html>The period of use of the software has expired. Enter the new key.<br>Contact e-mail: albtutanota@tutanota.com<html>");
@@ -53,32 +50,45 @@ public class LicenseWindow extends JFrame {
     }
 
     private void saveKey(String text) {
-        String zipFilePath = BaseConstants.getInstance().getZipFileSQL();
-        FileHeader fHeader=null;
-        ZipFile zipFile =null;
-        try {
-            zipFile=new ZipFile(zipFilePath);
-            if (zipFile.isEncrypted()){
-                zipFile.setPassword(BaseConstants.getInstance().getZipPsw());
-            }
-            fHeader =zipFile.getFileHeader("liesence.txt");
-            zipFile.removeFile(fHeader);
+        if (BaseConstants.isIsZip()) {
+            String zipFilePath = BaseConstants.getInstance().getZipFileSQL();
+            FileHeader fHeader = null;
+            ZipFile zipFile = null;
+            try {
+                zipFile = new ZipFile(zipFilePath);
+                if (zipFile.isEncrypted()) {
+                    zipFile.setPassword(BaseConstants.getInstance().getZipPsw());
+                }
+                fHeader = zipFile.getFileHeader("liesence.txt");
+                zipFile.removeFile(fHeader);
 
-        } catch (ZipException e) {
-            e.printStackTrace();
+            } catch (ZipException e) {
+                e.printStackTrace();
+            }
+            try (InputStream is = new ByteArrayInputStream(text.getBytes("utf-8"))) {
+                ZipParameters zp = new ZipParameters();
+                zp.setSourceExternalStream(true);
+                zp.setFileNameInZip(fHeader.getFileName());
+                zp.setPassword(BaseConstants.getInstance().getZipPsw());
+                zipFile.addStream(is, zp);
+            } catch (UnsupportedEncodingException e) {
+                e.printStackTrace();
+            } catch (IOException e) {
+                e.printStackTrace();
+            } catch (ZipException e) {
+                e.printStackTrace();
+            }
         }
-        try (InputStream is = new ByteArrayInputStream(text.getBytes("utf-8"))) {
-            ZipParameters zp = new ZipParameters();
-            zp.setSourceExternalStream(true);
-            zp.setFileNameInZip(fHeader.getFileName());
-            zp.setPassword(BaseConstants.getInstance().getZipPsw());
-            zipFile.addStream(is, zp);
-        } catch (UnsupportedEncodingException e) {
-            e.printStackTrace();
-        } catch (IOException e) {
-            e.printStackTrace();
-        } catch (ZipException e) {
-            e.printStackTrace();
+        else{
+            try (BufferedWriter bfw = new BufferedWriter(new FileWriter(BaseConstants.getInstance().getPathSQL() + "\\liesence.txt")))
+            {
+                bfw.write(text);
+            } catch (FileNotFoundException e) {
+                e.printStackTrace();
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
+
         }
     }
